functions/requirements.txt
# Function dependencies
google-ads==22.0.0
google-api-python-client==2.102.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
oauth2client==4.1.3
pillow==10.0.1
functions-framework==3.3.0
validators==0.22.0
